+++ internal/updates/votes.go
@@ -90,7 +90,7 @@ func (w *VoteWorker) LoadActive(ctx context.Context) error {
 
 func (w *VoteWorker) loop(ctx context.Context) error {
 	defer func(start time.Time) {
-		log.Info().Msgf("FQtx72k: votes loop took: %f seconds", time.Since(start).Seconds())
+		log.Info().Msgf("votes loop took: %f seconds", time.Since(start).Seconds())
 	}(time.Now())
 
 	ids, err := w.proposals.GetProposalForVotes(voteProposalLimit)
@@ -105,16 +105,13 @@ func (w *VoteWorker) loop(ctx context.Context) error {
 			snapshot.ListVotesWithProposalIDsFilter(id),
 		}
 
-		start := time.Now()
 		createdAfter := w.getLastVoteCreatedAtByProposal(id)
 		if !createdAfter.IsZero() {
 			opts = append(opts, snapshot.ListVotesCreatedAfter(createdAfter))
 		}
-		log.Info().Msgf("FQtx72k: getLastVoteCreatedAt[%s] took: %f seconds", id, time.Since(start).Seconds())
 
 		offset := 0
 		for {
-			start = time.Now()
 			votes, err := w.fetchVotes(ctx, offset, opts)
 			if errors.Is(err, snapshot.ErrTooManyRequests) {
 				log.Warn().Err(err).Msg("snapshot api limits are reached")
@@ -124,15 +121,12 @@ func (w *VoteWorker) loop(ctx context.Context) error {
 			if err != nil {
 				return err
 			}
-			log.Info().Msgf("FQtx72k: fetchVotes[%s] took: %f seconds", id, time.Since(start).Seconds())
 
 			log.Info().Int("count", len(votes)).Msg("fetched votes")
 
-			start = time.Now()
 			if err := w.processVotes(votes); err != nil {
 				return err
 			}
-			log.Info().Msgf("FQtx72k: processVotes[%s] took: %f seconds", id, time.Since(start).Seconds())
 
 			if len(votes) < votesPerRequest {
 				err := w.proposals.MarkVotesProcessed(id)